requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
lightgbm>=4.0
xgboost>=2.1
catboost>=1.2
optuna>=3.6
matplotlib>=3.9
plotly>=5.24
altair>=5.4
shap>=0.46
imbalanced-learn>=0.12
mlflow>=2.16
hydra-core>=1.3
pyyaml>=6.0
geopandas>=1.0
xarray>=2024.7.0
pyarrow>=17.0.0
pre-commit>=3.8
ruff>=0.6
black>=24.8
jupyterlab>=4.2